app.js modified and booklist.js
