+++ src/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
 import { Provider } from 'react-redux';
-import App from './components/app';
+import App from './containers/app';
 import store from './reducers/index';
+import './index.css';
 
 ReactDOM.render(
   <Provider store={store}>
